+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Read `release_notes.md` for commit level details.
 ## [Unreleased]
 
 ### Enhancements
+- Add screen record feature for Windows driver (https://github.com/appium/appium-windows-driver/pull/66)
+    - `#start_recording_screen`, `#stop_recording_screen`
 
 ### Bug fixes